# Using binary search instead of sequential search for rowData