# PR Details

<!--- Provide a general summary of your changes in the Title above -->
Using binary search instead of sequential search for rowData

## Description

<!--- Describe your changes in detail -->

## Related Issue
#1961 


<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->

## Motivation and Context

<!--- Why is this change required? What problem does it solve? -->

## How Has This Been Tested

<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- See how your change affects other areas of the code, etc. -->


Performance testing was conducted on an AMD64 WSL using Go's benchmark tool. The benchmarks were run the test mentioned in #1961 8 times . The following results were observed:

```
goos: linux
goarch: amd64
pkg: github.com/xuri/excelize/v2
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
               │ base.txt   │             new.txt               │
               │   sec/op   │   sec/op    vs base               │
GetFormulas-12   3.951 ± 4%   1.382 ± 3%  -65.02% (p=0.000 n=8)

               │  base.txt    │           new.txt             │
               │     B/op     │     B/op      vs base         │
GetFormulas-12   815.1Mi ± 0%   815.1Mi ± 0%  ~ (p=0.721 n=8)

               │  base.txt   │           new.txt            │
               │  allocs/op  │  allocs/op   vs base         │
GetFormulas-12   10.87M ± 0%   10.87M ± 0%  ~ (p=0.940 n=8)
(The version of the issue submitted is 27 seconds in the same environment)
```
code:
```

func BenchmarkGetFormulas(b *testing.B) {
	for i := 0; i < b.N; i++ {
		file, err := OpenFile("4600Formulas.xlsx")
		if err != nil {
			panic(err)
		}
		begin := time.Now()
		cc := 0
		for row := 1; row <= 4600; row++ { // 4600
			for col := 1; col <= 55; col++ { //55
				axis, _ := CoordinatesToCellName(col, row)
				str, err := file.GetCellFormula("Sheet1", axis)
				if err != nil {
					panic(err)
				}
				if str != "" {
					cc++
				}
			}
		}
		assert.Equal(b, 45990, cc)
		b.Logf("GetCellFormula time cost:%dms cc:%v", time.Since(begin).Milliseconds(), cc)
	}
}
```


## Types of changes

<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->

- [ ] Docs change / refactoring / dependency upgrade
- [ ] Bug fix (non-breaking change which fixes an issue)
- [x] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)

## Checklist

<!--- Go over all the following points, and put an `x` in all the boxes that apply. -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->

- [x] My code follows the code style of this project.
- [ ] My change requires a change to the documentation.
- [ ] I have updated the documentation accordingly.
- [x] I have read the **CONTRIBUTING** document.
- [ ] I have added tests to cover my changes.
- [x] All new and existing tests passed.