@shcabin 2025-05-11
this PR dependency that 'SheetData.Row[i].R ' is ordered.  
the first commit does not have this dependency and can reduce the time by 55% compared to the previous version (3.95seconds vs 1.762seconds).