@shcabin 2025-05-11
> This is awesome! Does these changes increase memory usage? If so, how much memories will be increased?

 no change in memory usage